refactor(utils): consolidate grouping functions into shared utility
